# swagger/swagger.yaml
openapi: 3.1.0
info:
  title: REST API
  description:
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/
paths:
  /auth/register:
    post:
      security:
        - bearerAuth: []
      tags:
        - Auth
      summary: User registration
      description: Creates a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "john.doe@example.com"
                password:
                  type: string
                  example: "MyStrongPassword123!"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email in use.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      security:
        - bearerAuth: []
      tags:
        - Auth
      summary: User login
      description: Login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "test@test.com"
                password:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: Login successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Validation Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The password or email are wrong"
  /auth/current:
    get:
      security:
        - bearerAuth: []
      tags:
        - Auth
      summary: Get current user
      description: Current
      responses:
        "200":
          description: Current user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: test@test.com
                  subscription:
                    type: string
                    example: started
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The password or email are wrong"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Not authorized"
  /auth/logout:
    post:
      security:
        - bearerAuth: []
      tags:
        - Auth
      summary: Logout current user
      description: Logout
      responses:
        "204":
          description: Logout successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Not authorized"
  /contacts:
    get:
      tags:
        - Contacts
      summary: Get all contacts
      operationId: listContacts
      responses:
        "200":
          description: List of contacts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      tags:
        - Contacts
      summary: Create a new contact
      operationId: addContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewContact"
      responses:
        "201":
          description: Contact created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Validation error.
  /contacts/{id}:
    get:
      tags:
        - Contacts
      summary: Get contact by ID
      operationId: getContactById
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: ID of the user to retrieve.
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not Found
    delete:
      tags:
        - Contacts
      summary: Delete contact by ID
      operationId: removeContact
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: ID of the user to delete.
      responses:
        "200":
          description: User was deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not Found
    put:
      tags:
        - Contacts
      summary: Update an existing contact
      operationId: updateContact
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: ID of the user to retrieve.
      requestBody:
        description: Contact object that needs to be updated
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContact"
      responses:
        "200":
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Validation error
        "404":
          description: Item not found
  /contacts/{id}/favorite:
    patch:
      tags:
        - Contacts
      summary: Update favorite status of a contact
      description: Sets or unsets the "favorite" flag for a specific contact.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: ID of the contact to update
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                favorite:
                  type: boolean
            example:
              favorite: true
      responses:
        "200":
          description: Favorite status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorite:
                    type: boolean
              example:
                favorite: true
        "400":
          description: Invalid request body
        "404":
          description: Contact not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT # optional, arbitrary value for documentation purposes"
  schemas:
    Contact:
      type: object
      properties:
        id:
          type: string
          example: AeHIrLTr6JkxGE6SN-0Rw
        name:
          type: string
          example: Allen Raymond
        email:
          type: string
          format: email
          example: nulla.ante@vestibul.co.uk
        phone:
          type: string
          example: (992) 914-3792
        favorite:
          type: boolean
          example: true
        createdAt:
          type: date
          example: 2025-11-21T18:03:56.764Z
        updatedAt:
          type: date
          example: 2025-11-21T18:03:56.764Z
    NewContact:
      type: object
      properties:
        name:
          type: string
          example: Allen Raymond
        email:
          type: string
          format: email
          example: nulla.ante@vestibul.co.uk
        phone:
          type: string
          example: (992) 914-3792
    UpdateContact:
      type: object
      properties:
        name:
          type: string
          example: Allen Raymond
        email:
          type: string
          format: email
          example: nulla.ante@vestibul.co.uk
        phone:
          type: string
          example: (992) 914-3792
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Email already exists"
        code:
          type: number
          example: 409